chatbot/requirements.txt
langchain==0.1.12
langchain-community>=0.0.28
langchain-openai>=0.0.8
langgraph==0.0.27
openai==1.12.0
python-dotenv==1.0.1
fastapi==0.110.0
uvicorn==0.27.1
pydantic==2.6.3
transformers==4.38.2
torch==2.2.1
nltk==3.8.1
python-multipart==0.0.9
httpx==0.27.0
pytest==8.0.2
pytest-asyncio==0.23.5
googlemaps==4.10.0
aiohttp==3.9.3
firebase-admin==6.4.0
twilio==8.12.0 